user permission added to apis

+++ src/core/services/index.ts
@@ -1,4 +1,4 @@
-import { AddProduct } from "../../interfaces";
+import { AddProduct, UpdateProduct } from "../../interfaces";
 import { ProductModel } from "../../database/schemas/products"
 
 
@@ -9,7 +9,8 @@ export async function addProduct(productObj: AddProduct): Promise<any> {
         description: productObj.description,
         amount: productObj.amount,
         pictures: productObj.pictures,
-        status: 1
+        status: 1,
+        user_id:productObj.user_id
     });
 
     let data = await msg.save()
@@ -29,9 +30,9 @@ export async function getProducts(): Promise<any> {
 
 }
 
-export async function deleteProduct(id: string): Promise<any> {
+export async function deleteProduct(id: string, userId: string): Promise<any> {
 
-    const data = await ProductModel.deleteOne({ _id: id })
+    const data = await ProductModel.deleteOne({ _id: id , user_id:userId})
     return {
         message: "Product deleted",
         data: data,
@@ -39,8 +40,9 @@ export async function deleteProduct(id: string): Promise<any> {
 
 }
 
-export async function updateProduct(id: string, productObj: AddProduct) {
-    var query = { '_id': id };
+export async function updateProduct(id: string, productObj: UpdateProduct) {
+    var query = { '_id': id ,'user_id':productObj.user_id};
+    delete productObj["user_id"]
 
     const data = await ProductModel.updateOne(query, productObj)
     return {
